backend/app/requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.215.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
aioredis==2.0.1
pandas==2.1.3
openpyxl==3.1.2
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.0.3
celery==5.3.4
redis==5.0.1
newsapi-python==0.2.7
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1